Restrucutre code so full-path imports work and we do not include a src folder in the project
