+++ fhir/DSTU2CapabilityStatement.go
@@ -2,8 +2,8 @@ package fhir
 
 import "encoding/xml"
 
-type DSTU3CapabilityStatement struct {
-	XMLName  xml.Name `xml:"CapabilityStatement"`
+type DSTU2CapabilityStatement struct {
+	XMLName  xml.Name `xml:"Conformance"`
 	Chardata string   `xml:",chardata"`
 	Xmlns    string   `xml:"xmlns,attr"`
 	ID       struct {
@@ -19,28 +19,13 @@ type DSTU3CapabilityStatement struct {
 		Div struct {
 			Text  string `xml:",chardata"`
 			Xmlns string `xml:"xmlns,attr"`
-			H2    string `xml:"h2"`
-			Div   struct {
+			P     []struct {
 				Text string `xml:",chardata"`
-				P    string `xml:"p"`
-			} `xml:"div"`
-			Table []struct {
-				Text string `xml:",chardata"`
-				Tr   []struct {
+				A    struct {
 					Text string `xml:",chardata"`
-					Td   []struct {
-						Text string `xml:",chardata"`
-						A    struct {
-							Text string `xml:",chardata"`
-							Href string `xml:"href,attr"`
-						} `xml:"a"`
-					} `xml:"td"`
-					Th []struct {
-						Text string `xml:",chardata"`
-						B    string `xml:"b"`
-					} `xml:"th"`
-				} `xml:"tr"`
-			} `xml:"table"`
+					Href string `xml:"href,attr"`
+				} `xml:"a"`
+			} `xml:"p"`
 		} `xml:"div"`
 	} `xml:"text"`
 	URL struct {
@@ -63,16 +48,16 @@ type DSTU3CapabilityStatement struct {
 		Text  string `xml:",chardata"`
 		Value string `xml:"value,attr"`
 	} `xml:"experimental"`
-	Date struct {
-		Text  string `xml:",chardata"`
-		Value string `xml:"value,attr"`
-	} `xml:"date"`
 	Publisher struct {
 		Text  string `xml:",chardata"`
 		Value string `xml:"value,attr"`
 	} `xml:"publisher"`
 	Contact struct {
-		Text    string `xml:",chardata"`
+		Text string `xml:",chardata"`
+		Name struct {
+			Text  string `xml:",chardata"`
+			Value string `xml:"value,attr"`
+		} `xml:"name"`
 		Telecom struct {
 			Text   string `xml:",chardata"`
 			System struct {
@@ -85,10 +70,22 @@ type DSTU3CapabilityStatement struct {
 			} `xml:"value"`
 		} `xml:"telecom"`
 	} `xml:"contact"`
+	Date struct {
+		Text  string `xml:",chardata"`
+		Value string `xml:"value,attr"`
+	} `xml:"date"`
 	Description struct {
 		Text  string `xml:",chardata"`
 		Value string `xml:"value,attr"`
 	} `xml:"description"`
+	Requirements struct {
+		Text  string `xml:",chardata"`
+		Value string `xml:"value,attr"`
+	} `xml:"requirements"`
+	Copyright struct {
+		Text  string `xml:",chardata"`
+		Value string `xml:"value,attr"`
+	} `xml:"copyright"`
 	Kind struct {
 		Text  string `xml:",chardata"`
 		Value string `xml:"value,attr"`
@@ -99,7 +96,26 @@ type DSTU3CapabilityStatement struct {
 			Text  string `xml:",chardata"`
 			Value string `xml:"value,attr"`
 		} `xml:"name"`
+		Version struct {
+			Text  string `xml:",chardata"`
+			Value string `xml:"value,attr"`
+		} `xml:"version"`
+		ReleaseDate struct {
+			Text  string `xml:",chardata"`
+			Value string `xml:"value,attr"`
+		} `xml:"releaseDate"`
 	} `xml:"software"`
+	Implementation struct {
+		Text        string `xml:",chardata"`
+		Description struct {
+			Text  string `xml:",chardata"`
+			Value string `xml:"value,attr"`
+		} `xml:"description"`
+		URL struct {
+			Text  string `xml:",chardata"`
+			Value string `xml:"value,attr"`
+		} `xml:"url"`
+	} `xml:"implementation"`
 	FhirVersion struct {
 		Text  string `xml:",chardata"`
 		Value string `xml:"value,attr"`
@@ -129,8 +145,8 @@ type DSTU3CapabilityStatement struct {
 				Value string `xml:"value,attr"`
 			} `xml:"cors"`
 			Service struct {
-				Chardata string `xml:",chardata"`
-				Coding   struct {
+				Text   string `xml:",chardata"`
+				Coding struct {
 					Text   string `xml:",chardata"`
 					System struct {
 						Text  string `xml:",chardata"`
@@ -140,22 +156,25 @@ type DSTU3CapabilityStatement struct {
 						Text  string `xml:",chardata"`
 						Value string `xml:"value,attr"`
 					} `xml:"code"`
-					Display struct {
-						Text  string `xml:",chardata"`
-						Value string `xml:"value,attr"`
-					} `xml:"display"`
 				} `xml:"coding"`
-				Text struct {
-					Text  string `xml:",chardata"`
-					Value string `xml:"value,attr"`
-				} `xml:"text"`
 			} `xml:"service"`
 			Description struct {
 				Text  string `xml:",chardata"`
 				Value string `xml:"value,attr"`
 			} `xml:"description"`
+			Certificate struct {
+				Text string `xml:",chardata"`
+				Type struct {
+					Text  string `xml:",chardata"`
+					Value string `xml:"value,attr"`
+				} `xml:"type"`
+				Blob struct {
+					Text  string `xml:",chardata"`
+					Value string `xml:"value,attr"`
+				} `xml:"blob"`
+			} `xml:"certificate"`
 		} `xml:"security"`
-		Resource []struct {
+		Resource struct {
 			Text string `xml:",chardata"`
 			Type struct {
 				Text  string `xml:",chardata"`
@@ -179,6 +198,18 @@ type DSTU3CapabilityStatement struct {
 					Value string `xml:"value,attr"`
 				} `xml:"documentation"`
 			} `xml:"interaction"`
+			Versioning struct {
+				Text  string `xml:",chardata"`
+				Value string `xml:"value,attr"`
+			} `xml:"versioning"`
+			ReadHistory struct {
+				Text  string `xml:",chardata"`
+				Value string `xml:"value,attr"`
+			} `xml:"readHistory"`
+			UpdateCreate struct {
+				Text  string `xml:",chardata"`
+				Value string `xml:"value,attr"`
+			} `xml:"updateCreate"`
 			ConditionalCreate struct {
 				Text  string `xml:",chardata"`
 				Value string `xml:"value,attr"`
@@ -191,15 +222,11 @@ type DSTU3CapabilityStatement struct {
 				Text  string `xml:",chardata"`
 				Value string `xml:"value,attr"`
 			} `xml:"conditionalDelete"`
-			ReferencePolicy []struct {
-				Text  string `xml:",chardata"`
-				Value string `xml:"value,attr"`
-			} `xml:"referencePolicy"`
-			SearchInclude []struct {
+			SearchInclude struct {
 				Text  string `xml:",chardata"`
 				Value string `xml:"value,attr"`
 			} `xml:"searchInclude"`
-			SearchRevInclude []struct {
+			SearchRevInclude struct {
 				Text  string `xml:",chardata"`
 				Value string `xml:"value,attr"`
 			} `xml:"searchRevInclude"`
@@ -221,6 +248,18 @@ type DSTU3CapabilityStatement struct {
 					Text  string `xml:",chardata"`
 					Value string `xml:"value,attr"`
 				} `xml:"documentation"`
+				Modifier struct {
+					Text  string `xml:",chardata"`
+					Value string `xml:"value,attr"`
+				} `xml:"modifier"`
+				Target struct {
+					Text  string `xml:",chardata"`
+					Value string `xml:"value,attr"`
+				} `xml:"target"`
+				Chain []struct {
+					Text  string `xml:",chardata"`
+					Value string `xml:"value,attr"`
+				} `xml:"chain"`
 			} `xml:"searchParam"`
 		} `xml:"resource"`
 		Interaction []struct {
@@ -229,43 +268,101 @@ type DSTU3CapabilityStatement struct {
 				Text  string `xml:",chardata"`
 				Value string `xml:"value,attr"`
 			} `xml:"code"`
-			Documentation struct {
-				Text  string `xml:",chardata"`
-				Value string `xml:"value,attr"`
-			} `xml:"documentation"`
 		} `xml:"interaction"`
-		SearchParam []struct {
+		Compartment struct {
+			Text  string `xml:",chardata"`
+			Value string `xml:"value,attr"`
+		} `xml:"compartment"`
+	} `xml:"rest"`
+	Messaging struct {
+		Text     string `xml:",chardata"`
+		Endpoint struct {
+			Text     string `xml:",chardata"`
+			Protocol struct {
+				Text   string `xml:",chardata"`
+				System struct {
+					Text  string `xml:",chardata"`
+					Value string `xml:"value,attr"`
+				} `xml:"system"`
+				Code struct {
+					Text  string `xml:",chardata"`
+					Value string `xml:"value,attr"`
+				} `xml:"code"`
+			} `xml:"protocol"`
+			Address struct {
+				Text  string `xml:",chardata"`
+				Value string `xml:"value,attr"`
+			} `xml:"address"`
+		} `xml:"endpoint"`
+		ReliableCache struct {
+			Text  string `xml:",chardata"`
+			Value string `xml:"value,attr"`
+		} `xml:"reliableCache"`
+		Documentation struct {
+			Text  string `xml:",chardata"`
+			Value string `xml:"value,attr"`
+		} `xml:"documentation"`
+		Event struct {
 			Text string `xml:",chardata"`
-			Name struct {
+			Code struct {
+				Text   string `xml:",chardata"`
+				System struct {
+					Text  string `xml:",chardata"`
+					Value string `xml:"value,attr"`
+				} `xml:"system"`
+				Code struct {
+					Text  string `xml:",chardata"`
+					Value string `xml:"value,attr"`
+				} `xml:"code"`
+			} `xml:"code"`
+			Category struct {
 				Text  string `xml:",chardata"`
 				Value string `xml:"value,attr"`
-			} `xml:"name"`
-			Definition struct {
+			} `xml:"category"`
+			Mode struct {
 				Text  string `xml:",chardata"`
 				Value string `xml:"value,attr"`
-			} `xml:"definition"`
-			Type struct {
+			} `xml:"mode"`
+			Focus struct {
 				Text  string `xml:",chardata"`
 				Value string `xml:"value,attr"`
-			} `xml:"type"`
-			Documentation struct {
-				Text  string `xml:",chardata"`
-				Value string `xml:"value,attr"`
-			} `xml:"documentation"`
-		} `xml:"searchParam"`
-		Operation []struct {
-			Text string `xml:",chardata"`
-			Name struct {
-				Text  string `xml:",chardata"`
-				Value string `xml:"value,attr"`
-			} `xml:"name"`
-			Definition struct {
+			} `xml:"focus"`
+			Request struct {
 				Text      string `xml:",chardata"`
 				Reference struct {
 					Text  string `xml:",chardata"`
 					Value string `xml:"value,attr"`
 				} `xml:"reference"`
-			} `xml:"definition"`
-		} `xml:"operation"`
-	} `xml:"rest"`
-} 
+			} `xml:"request"`
+			Response struct {
+				Text      string `xml:",chardata"`
+				Reference struct {
+					Text  string `xml:",chardata"`
+					Value string `xml:"value,attr"`
+				} `xml:"reference"`
+			} `xml:"response"`
+			Documentation struct {
+				Text  string `xml:",chardata"`
+				Value string `xml:"value,attr"`
+			} `xml:"documentation"`
+		} `xml:"event"`
+	} `xml:"messaging"`
+	Document struct {
+		Text string `xml:",chardata"`
+		Mode struct {
+			Text  string `xml:",chardata"`
+			Value string `xml:"value,attr"`
+		} `xml:"mode"`
+		Documentation struct {
+			Text  string `xml:",chardata"`
+			Value string `xml:"value,attr"`
+		} `xml:"documentation"`
+		Profile struct {
+			Text      string `xml:",chardata"`
+			Reference struct {
+				Text  string `xml:",chardata"`
+				Value string `xml:"value,attr"`
+			} `xml:"reference"`
+		} `xml:"profile"`
+	} `xml:"document"`
+}
