+++ fhir/testdata/DSTU2CapabilityStatement.xml
