+++ fhir/testdata/DSTU3CapabilityStatement.xml
